Maybe fix error

+++ src/ui.c
@@ -18,7 +18,7 @@ int uiGetActive(void)
 int GetActiveUIPointer(char UI)
 {
 	int UI_POINTER_LIST = 0x011C7064;
-	int Pointer = (*(u32*)((u32)UI_POINTER_LIST + (UI * 0x4)));
+	int Pointer = *(u32*)(UI_POINTER_LIST + (UI * 0x4));
 	int ActiveUIPointer = (*(u32*)0x011C7108);
 	if (ActiveUIPointer == Pointer)
 	{
